requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.12
django-ckeditor==5.4.0
django-colorfield==0.1.14
django-extensions==2.0.6
django-faker==0.2
django-filer==1.3.2
django-filter==1.1.0
django-hvad==1.8.0
django-js-asset==1.0.0
django-mptt==0.8.7
django-parler==1.9.2
django-polymorphic==1.3
django-redis==4.9.0
django-rest-swagger==2.1.2
django-suit==0.2.26
djangorestframework==3.7.7
easy-thumbnails==2.4.2
et-xmlfile==1.0.1
fake-factory==9999.9.9
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
lml==0.0.1
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.2
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.0
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pyexcel==0.5.8
pyexcel-io==0.5.6
pyexcel-xlsx==0.5.5
python-dateutil==2.7.0
python-magic==0.4.15
pytz==2017.3
PyYAML==3.12
raven==6.7.0
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
six==1.11.0
solid-i18n==1.4.1
tablib==0.12.1
texttable==1.2.1
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
